requirements.txt
# Web Framework
flask==3.1.0

# Environment Management
python-dotenv==1.1.0

# Data Processing (Phase 1+)
pandas==2.2.3

# Hugging Face Dataset (Phase 1+)
datasets==3.5.0

# Fuzzy Matching (Phase 2+)
rapidfuzz==3.12.2

# LLM Integration (Phase 3+)
openai==1.78.0
google-generativeai==0.8.5

# Utilities
requests==2.32.3
groq==0.11.0
fastapi
uvicorn
